Introduce product data importer
Why:
We would like to have a dedicated service object to handle the logic of
importing csv files for product data.

This commit:
Introduces a ProductDataImporter model and moves the import
functionality from Product.import to
ProductDataImporter#import
