+++ app/controllers/imports_controller.rb
@@ -1,6 +1,7 @@
 class ImportsController < ApplicationController
   def create
-    Product.import(params[:file].path)
+    importer = ProductDataImporter.new(params[:file].path)
+    importer.import
     redirect_to products_path, notice: "Succesfully imported"
   end
 end
