+++ app/models/product.rb
@@ -2,13 +2,4 @@ require "csv"
 
 class Product < ApplicationRecord
   validates :name, presence: true
-
-  def self.import(file_path)
-    CSV.foreach(file_path, headers: true) do |row|
-      data = row.to_h
-      data["active"] = data["active"] == "true"
-      product = new(data)
-      product.save
-    end
-  end
 end
